Scoped Pool for scheduling non 'static futures
+ Provide `juliex::scope::Pool` that integrates with
  `crossbeam::scope` to allow easy spawning of not `'static`
  futures.

+ Refactor `Task` structure into separate module for usage
  across different executors

+ Replace `TaskQueue` as synonym for `channel::Sender` to
  better separate from queue receivers (the worker threads)
