+++ src/task.rs
@@ -1,0 +1,185 @@
+use std::cell::UnsafeCell;
+use std::fmt;
+use std::future::Future;
+use std::mem::{forget, ManuallyDrop};
+use std::sync::{
+    atomic::{AtomicUsize, Ordering::SeqCst},
+    Arc,
+};
+use std::task::{Context, Poll, RawWaker, RawWakerVTable, Waker};
+use crossbeam::channel;
+use futures::future::BoxFuture;
+use futures::prelude::*;
+
+pub type TaskQueue = channel::Sender<Task>;
+
+#[derive(Clone, Debug)]
+#[repr(transparent)]
+pub struct Task(Arc<AtomicFuture>);
+
+struct AtomicFuture {
+    queue: Arc<TaskQueue>,
+    status: AtomicUsize,
+    future: UnsafeCell<BoxFuture<'static, ()>>,
+}
+
+impl fmt::Debug for AtomicFuture {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        "AtomicFuture".fmt(f)
+    }
+}
+
+unsafe impl Send for AtomicFuture {}
+unsafe impl Sync for AtomicFuture {}
+
+const WAITING: usize = 0; // --> POLLING
+const POLLING: usize = 1; // --> WAITING, REPOLL, or COMPLETE
+const REPOLL: usize = 2; // --> POLLING
+const COMPLETE: usize = 3; // No transitions out
+
+impl Task {
+    #[inline]
+    pub fn new<F: Future<Output = ()> + Send + 'static>(future: F, queue: Arc<TaskQueue>) -> Task {
+        let future: Arc<AtomicFuture> = Arc::new(AtomicFuture {
+            queue,
+            status: AtomicUsize::new(WAITING),
+            future: UnsafeCell::new(future.boxed()),
+        });
+        let future: *const AtomicFuture = Arc::into_raw(future) as *const AtomicFuture;
+        unsafe { task(future) }
+    }
+
+    #[inline]
+    pub fn new_boxed(future: BoxFuture<'static, ()>, queue: Arc<TaskQueue>) -> Task {
+        let future: Arc<AtomicFuture> = Arc::new(AtomicFuture {
+            queue,
+            status: AtomicUsize::new(WAITING),
+            future: UnsafeCell::new(future),
+        });
+        let future: *const AtomicFuture = Arc::into_raw(future) as *const AtomicFuture;
+        unsafe { task(future) }
+    }
+
+    #[inline]
+    pub unsafe fn poll(self) {
+        self.0.status.store(POLLING, SeqCst);
+        let waker = ManuallyDrop::new(waker(&*self.0));
+        let mut cx = Context::from_waker(&waker);
+        loop {
+            if let Poll::Ready(_) = (&mut *self.0.future.get()).poll_unpin(&mut cx) {
+                break self.0.status.store(COMPLETE, SeqCst);
+            }
+            match self
+                .0
+                .status
+                .compare_exchange(POLLING, WAITING, SeqCst, SeqCst)
+            {
+                Ok(_) => break,
+                Err(_) => self.0.status.store(POLLING, SeqCst),
+            }
+        }
+    }
+}
+
+#[inline]
+unsafe fn waker(task: *const AtomicFuture) -> Waker {
+    Waker::from_raw(RawWaker::new(
+        task as *const (),
+        &RawWakerVTable::new(clone_raw, wake_raw, wake_ref_raw, drop_raw),
+    ))
+}
+
+#[inline]
+unsafe fn clone_raw(this: *const ()) -> RawWaker {
+    let task = clone_task(this as *const AtomicFuture);
+    RawWaker::new(
+        Arc::into_raw(task.0) as *const (),
+        &RawWakerVTable::new(clone_raw, wake_raw, wake_ref_raw, drop_raw),
+    )
+}
+
+#[inline]
+unsafe fn drop_raw(this: *const ()) {
+    drop(task(this as *const AtomicFuture))
+}
+
+#[inline]
+unsafe fn wake_raw(this: *const ()) {
+    let task = task(this as *const AtomicFuture);
+    let mut status = task.0.status.load(SeqCst);
+    loop {
+        match status {
+            WAITING => {
+                match task
+                    .0
+                    .status
+                    .compare_exchange(WAITING, POLLING, SeqCst, SeqCst)
+                {
+                    Ok(_) => {
+                        task.0.queue.send(clone_task(&*task.0)).unwrap();
+                        break;
+                    }
+                    Err(cur) => status = cur,
+                }
+            }
+            POLLING => {
+                match task
+                    .0
+                    .status
+                    .compare_exchange(POLLING, REPOLL, SeqCst, SeqCst)
+                {
+                    Ok(_) => break,
+                    Err(cur) => status = cur,
+                }
+            }
+            _ => break,
+        }
+    }
+}
+
+#[inline]
+unsafe fn wake_ref_raw(this: *const ()) {
+    let task = ManuallyDrop::new(task(this as *const AtomicFuture));
+    let mut status = task.0.status.load(SeqCst);
+    loop {
+        match status {
+            WAITING => {
+                match task
+                    .0
+                    .status
+                    .compare_exchange(WAITING, POLLING, SeqCst, SeqCst)
+                {
+                    Ok(_) => {
+                        task.0.queue.send(clone_task(&*task.0)).unwrap();
+                        break;
+                    }
+                    Err(cur) => status = cur,
+                }
+            }
+            POLLING => {
+                match task
+                    .0
+                    .status
+                    .compare_exchange(POLLING, REPOLL, SeqCst, SeqCst)
+                {
+                    Ok(_) => break,
+                    Err(cur) => status = cur,
+                }
+            }
+            _ => break,
+        }
+    }
+}
+
+#[inline]
+unsafe fn task(future: *const AtomicFuture) -> Task {
+    Task(Arc::from_raw(future))
+}
+
+#[inline]
+unsafe fn clone_task(future: *const AtomicFuture) -> Task {
+    let task = task(future);
+    forget(task.clone());
+    task
+}
